+++ crates/ferrmion-core/src/optimise/ternarytree/mod.rs
@@ -1,4 +1,6 @@
 mod hatt;
 pub use hatt::*;
+mod term_store;
+pub use term_store::*;
 mod topphatt;
 pub use topphatt::*;